Python/embedclipimgasbase64md/requirements.txt
pillow
pyperclip
pyinstaller